+++ src/shape/ellipse.rs
@@ -333,12 +333,7 @@ impl<T: Num> Ellipse<T> {
     /// Returns the bounding [Rect] of the ellipse.
     #[inline]
     pub fn bounding_rect(&self) -> Rect<T> {
-        rect![
-            self.left(),
-            self.top(),
-            self.width() + T::one(),
-            self.height() + T::one()
-        ]
+        rect![self.left(), self.top(), self.width(), self.height()]
     }
 
     /// Returns `Ellipse` as a [Vec].